+++ src/utils/EnvironmentUtils.js
